[GD-71] refactor : 도메인 레포지토리 생성 및 ddl 변경
