+++ src/main/java/com/dokev/gold_dduck/event/domain/EventProgressStatus.java
@@ -1,4 +1,4 @@
-package com.dokev.gold_dduck.event;
+package com.dokev.gold_dduck.event.domain;
 
 public enum EventProgressStatus {
     READY, RUNNING, CLOSED
