+++ src/main/java/com/dokev/gold_dduck/gift/domain/GiftType.java
@@ -1,4 +1,4 @@
-package com.dokev.gold_dduck.gift;
+package com.dokev.gold_dduck.gift.domain;
 
 public enum GiftType {
     URL, TEXT, IMAGE
