+++ src/main/java/com/dokev/gold_dduck/gift/repository/GiftItemRepository.java
@@ -1,0 +1,7 @@
+package com.dokev.gold_dduck.gift.repository;
+
+import com.dokev.gold_dduck.gift.domain.GiftItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GiftItemRepository extends JpaRepository<GiftItem, Long> {
+}
